Initial Commit
this is a small GUI to calculate load averages on Windows and will update
them every 2 seconds. The GUI has issues quitting after a while (probably
a race condition somewhere).
This load time calculation also includes Disk I/O in an attempt to
replicate how Linux calculates loads but I don't currently know of a way
to get how many threads are waiting on Disk I/O instead of how many
objects are wanting to write to the disk (eg, one thread can have 3
objects to write to disk but this program will count it as 4 objects)